Create Dice Roll Game.py
The code depicts two inputs as predicted by the computer as Player 1 and Player 2 inputs. 
If the number predicted for player 1 is greater than that predicted for player 2, player 1 scores and vice - versa.
This continues for 10 rounds, and the final result is predicted according to the scores at the end of round 10.